refactor: s3 관련 Type 클래스 패키지 이동
- ImgType을 s3
