+++ src/main/java/com/example/solidconnection/community/post/service/PostCommandService.java
@@ -16,7 +16,7 @@ import com.example.solidconnection.s3.UploadedFileUrlResponse;
 import com.example.solidconnection.service.RedisService;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
-import com.example.solidconnection.type.ImgType;
+import com.example.solidconnection.s3.ImgType;
 import com.example.solidconnection.community.post.domain.PostCategory;
 import com.example.solidconnection.util.RedisUtils;
 import lombok.RequiredArgsConstructor;
