+++ src/test/java/com/example/solidconnection/siteuser/service/MyPageServiceTest.java
@@ -9,7 +9,7 @@ import com.example.solidconnection.siteuser.dto.NicknameUpdateRequest;
 import com.example.solidconnection.siteuser.repository.LikedUniversityRepository;
 import com.example.solidconnection.siteuser.repository.SiteUserRepository;
 import com.example.solidconnection.support.integration.BaseIntegrationTest;
-import com.example.solidconnection.type.ImgType;
+import com.example.solidconnection.s3.ImgType;
 import com.example.solidconnection.type.PreparationStatus;
 import com.example.solidconnection.type.Role;
 import com.example.solidconnection.university.domain.LikedUniversity;
